refactor: 태그 결정 책임을 enum 클래스로 위임
