+++ tamago-core/src/main/kotlin/tamago/server/core/letter/application/event/LetterEventListener.kt
@@ -6,11 +6,11 @@ import org.springframework.transaction.annotation.Propagation
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.transaction.event.TransactionalEventListener
 import tamago.server.core.letter.application.service.LetterQueryService
-import tamago.server.core.letter.application.service.LetterTagResolver
 import tamago.server.core.letter.application.service.UserLetterCommandService
 import tamago.server.core.letter.application.service.UserLetterQueryService
 import tamago.server.core.letter.domain.aggregate.UserLetter
 import tamago.server.core.letter.domain.enum.LetterStatus
+import tamago.server.core.letter.domain.enum.LetterTag
 import tamago.server.core.running.domain.event.RunningCompletedEvent
 
 private val logger = KotlinLogging.logger {}
@@ -18,7 +18,6 @@ private val logger = KotlinLogging.logger {}
 @Component
 class LetterEventListener(
     private val letterQueryService: LetterQueryService,
-    private val letterTagResolver: LetterTagResolver,
     private val userLetterQueryService: UserLetterQueryService,
     private val userLetterCommandService: UserLetterCommandService,
 ) {
@@ -34,7 +33,7 @@ class LetterEventListener(
             }
 
             // 러닝 상태에 맞는 태그 결정하고 다음 편지 세팅
-            val tag = letterTagResolver.resolve(event.isFirstRun, event.streakDays, event.inactiveDays)
+            val tag = LetterTag.resolve(event.isFirstRun, event.streakDays, event.inactiveDays)
             val selectedLetter = letterQueryService.getRandomTemplateByTag(tag)
 
             // 러닝 시작 시간 기준 23시간 후로 편지 발송 시간 설정
